added a class that stores passwords, usernames, and keys to text files
When an object is created the user provides the name of a text file, password, username, and key. The text file is then created and the contents are formatted and wrote to it. This class has a second constructor that allows the object to be made for the purpose of updating either the password, username, or key. There is a main method in the class that showcases how everything works.  This class also has a method that can print the contents of the file.