Functional Login Screen.
Created a new passwordEncryptionForExistingLogin class, contains a EncryptedLoginPassword method which encrypted a usernameinput with a key input.

Created A IfAccountExists class, parseFileForUsername method parses file if it contains the usernamesearch.

Created UsernameEncryption class, EncryptedUsername method encrypts a username with the username encryption key, all account usernames use the same encryption key, as the username is used as a ID value in the txt file and MySQL database.

LoginScreen class starts with the Login Screen page which checks if the username and password are valid and exists in the txt file. The bottom of the Login Screen page contains a hyperlink that redirect the user to the Create An Account Screen page. This page allows the user to create a new account. New username and password, with the key for the password associated with the account are stored on the txt file.

Modified StoredCredentials to append to the txt file with a newline.
Fixed errors with methods in LoginAuthenticationForTextFile class.