Added password change program.
Receives the stored username from the dashboard, old password and new password from user input.
Checks that the username and old password are valid by calling authenticationfortextfile();.
Parses the storedCredentials.txt file to recreate the line as a string.
Creates a string newline with the current username, new updated password, and current key.
Uses RandomAccessFile to replace the contents of the old line with the new line string.

Additionally I added a label on the dashboard that displays the currently logged in user.