Added change password feature to the dashboard
Modified the dashboard.fxml file and the dashboard controller so that the user can click a button that opens the changepassword.fxml file. In this page the user enters their username, old password, and new password then clicks a button which changes their login password. There is also another button that lets the user navigate back to the dashboard.