Added Security check fxml and controller
Created a new fxml file that makes the user enter the answers to their security questions before they can proceed to the next page. Currently its set up to go to the Change Password fxml, however, the code can easily be copied and changed to take you somewhere else. NOTE_1: had to fix a bug that was caused by the parsefile method in the login authentication. It was returning a byte array error bc the substring line that picks out the key didnt have an endpoint. NOTE_2: removed storeCredentials ability to encrypt the answers for the security check, so now you have to feed storeCredentials the answers in their encrypted format.